2003-01-13  Roland McGrath  <roland@redhat.com>
	* desc.c (printflock64): Fix ADDR argument type.
	From Anton Blanchard <anton@samba.org>.

+++ desc.c
@@ -212,7 +212,7 @@ int getlk;
 static void
 printflock64(tcp, addr, getlk)
 struct tcb *tcp;
-int addr;
+long addr;
 int getlk;
 {
 	struct flock64 fl;
